+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.68
 ### is.codion.swing.framework.ui
 - EntitySearchFieldPanel.Builder return types fixed.
+- EntityEditComponentPanel.EditorComponent.set(ComponentValueBuilder) added, setComponentBuilder() removed.
 
 ## 0.18.67
 ### is.codion.common.reactive